pyproject.toml
[project]
name = "slotmachine"
version = "0.1.0"
description = "MCP-based ClaudeCode plugin — Obsidian vault를 GraphDB에 연동하는 Second Brain"
requires-python = ">=3.11"
dependencies = [
    # MCP 서버 프레임워크
    "fastmcp>=2.0",
    # GraphDB
    "neo4j>=5.0",
    # Git 연동
    "gitpython>=3.1",
    # Markdown 파싱
    "python-frontmatter>=1.1",
    "mistletoe>=1.3",
    # 설정 관리
    "pydantic-settings>=2.0",
    # HTTP 클라이언트 (Jina 등 REST 기반 프로바이더)
    "httpx>=0.27",
    # 임베딩 프로바이더 SDK
    "voyageai>=0.3",        # Voyage AI
    "openai>=1.0",          # OpenAI
    "google-genai>=1.0",    # Gemini
    "ollama>=0.3",          # Ollama (로컬)
    # Webhook 서버
    "fastapi>=0.111",
    "uvicorn>=0.30",
    # 유틸리티
    "tqdm>=4.66",           # 진행률 표시
    "tenacity>=8.0",        # 재시도 로직
    "numpy>=2.4.3",
]

[project.scripts]
slotmachine = "slotmachine.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",   # httpx mock
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
